refactor: split onboard starting point render
