+++ crates/daemon/src/onboard_cli.rs
@@ -81,12 +81,15 @@ use self::select_support::*;
 mod screen_spec_support;
 
 use self::screen_spec_support::*;
+#[path = "onboard_starting_point_render.rs"]
+mod starting_point_render_support;
 #[path = "onboard_starting_point.rs"]
 mod starting_point_support;
 use self::starting_point_support::*;
 pub use self::starting_point_support::{
-    detect_import_starting_config_with_channel_readiness, should_offer_current_setup_shortcut,
-    should_offer_detected_setup_shortcut, validate_non_interactive_risk_gate,
+    collect_import_candidates_with_paths, detect_import_starting_config_with_channel_readiness,
+    should_offer_current_setup_shortcut, should_offer_detected_setup_shortcut,
+    validate_non_interactive_risk_gate,
 };
 #[path = "onboard_guided_config.rs"]
 mod guided_config_support;
@@ -100,8 +103,7 @@ mod onboard_cli_render;
 #[path = "onboard_review_render.rs"]
 mod onboard_review_render;
 pub use self::onboard_cli_render::{
-    append_escape_cancel_hint, collect_import_candidates_with_paths,
-    render_api_key_env_selection_screen_lines,
+    append_escape_cancel_hint, render_api_key_env_selection_screen_lines,
     render_api_key_env_selection_screen_lines_with_default,
     render_continue_current_setup_screen_lines, render_continue_detected_setup_screen_lines,
     render_current_setup_write_confirmation_screen_lines, render_default_choice_footer_line,
@@ -109,8 +111,7 @@ pub use self::onboard_cli_render::{
     render_existing_config_write_screen_lines, render_model_selection_screen_lines,
     render_model_selection_screen_lines_with_default, render_onboard_entry_screen_lines,
     render_onboarding_risk_screen_lines, render_provider_selection_screen_lines,
-    render_single_detected_setup_preview_screen_lines,
-    render_starting_point_selection_screen_lines, render_system_prompt_selection_screen_lines,
+    render_system_prompt_selection_screen_lines,
     render_system_prompt_selection_screen_lines_with_default,
     render_write_confirmation_screen_lines,
 };
@@ -122,14 +123,12 @@ use self::onboard_cli_render::{
     render_onboard_shortcut_header_lines_with_style, render_prompt_with_default_text,
     render_provider_selection_header_lines, render_system_prompt_selection_screen_lines_with_style,
     render_web_search_credential_selection_screen_lines_with_style, screen_subtitle,
-    select_interactive_import_starting_config, start_fresh_starting_point_detail_lines,
-    summarize_starting_point_detail_lines, tui_header_style,
+    tui_header_style,
 };
 #[cfg(test)]
 use self::onboard_cli_render::{
     render_onboard_option_lines, render_onboard_option_prefix,
     render_onboard_shortcut_screen_lines_with_style,
-    render_starting_point_selection_header_lines_with_style,
 };
 #[cfg(test)]
 use self::onboard_review_render::provider_matches_for_review;
@@ -141,6 +140,11 @@ pub use self::onboard_review_render::{
     render_current_setup_review_lines_with_guidance,
     render_detected_setup_review_lines_with_guidance, render_onboard_review_lines_with_guidance,
     summarize_prompt_addendum, summarize_prompt_mode, summarize_provider_credential,
+};
+#[cfg(test)]
+use self::starting_point_render_support::render_starting_point_selection_header_lines_with_style;
+pub use self::starting_point_render_support::{
+    render_single_detected_setup_preview_screen_lines, render_starting_point_selection_screen_lines,
 };
 pub use crate::onboard_finalize::{
     OnboardingAction, OnboardingActionKind, OnboardingChannelSurfaceSummary,
